backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.10.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
cryptography==43.0.1
slowapi==0.1.9
psutil==5.9.8

